updated: Rethinking Data Format
Having in mind that eventually all will be Serialized as JSON data,
we are using `BTreeMaps`, and `BTreeSet`s to be consistant with the
original data. These will guranntee that the ordering of these elements
will always be the same. This doesn't help the voting modules perse, but
more importantly the order of each entry should be consistant when
calculating the hash.